// src/interface.ts
export enum LiveRoomIsShowEnum {
  /** 显示 */
  yes,
  /** 不显示 */
  no,
}
// ...
export enum FormTypeEnum {
  'input' = 'input',
  'password' = 'password',
  'number' = 'number',
  'select' = 'select',
  'radio' = 'radio',
  'checkbox' = 'checkbox',
  'markdown' = 'markdown',
  'switch' = 'switch',
  'upload' = 'upload',
  'treeSelect' = 'treeSelect',
  'datePickerRange' = 'datePickerRange',
  'datePickerDatetime' = 'datePickerDatetime',
}
// ...
export enum bankTypeEnum {
  BANK = 1,
  VIRTUAL = 2,
}
// ...
export interface IBank {
  id?: number;
  bank_name?: string;
  is_domestic?: number;
  protocol_type_list?: string[];
  logo?: string;
  type?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}